refactor(client): extract message handling to reduce file length
Move text/binary dispatch, protocol message handlers, and audio chunk
processing (~420 lines) into SendspinClient+MessageHandling.swift.

SendspinClient.swift drops from 1115 to 696 lines, well under the
900-line SwiftLint warning threshold. Properties accessed by the
extension changed from private to internal (same-module visibility).
