+++ Sources/SendspinKit/Client/SendspinClient.swift
@@ -10,29 +10,33 @@ import os
 @MainActor
 public final class SendspinClient {
     // Configuration
-    private let clientId: String
-    private let name: String
-    private let roles: Set<VersionedRole>
-    private let playerConfig: PlayerConfiguration?
-    private let artworkConfig: ArtworkConfiguration?
+    let clientId: String
+    let name: String
+    let roles: Set<VersionedRole>
+    let playerConfig: PlayerConfiguration?
+    let artworkConfig: ArtworkConfiguration?
     /// Resolved volume capabilities and control implementation
-    private let volumeCapabilities: VolumeCapabilities
-    private let volumeControl: VolumeControl
+    let volumeCapabilities: VolumeCapabilities
+    let volumeControl: VolumeControl
+
+    // All properties below are `internal` (not `private`) so that extension files
+    // in the same module (SendspinClient+Commands.swift, SendspinClient+MessageHandling.swift)
+    // can access them. They are NOT part of the public API — only `public` properties are.
 
     /// State
-    public private(set) var connectionState: ConnectionState = .disconnected
+    public internal(set) var connectionState: ConnectionState = .disconnected
     /// The audio format currently being streamed by the server, or nil if no stream is active.
-    public private(set) var currentStreamFormat: AudioFormatSpec?
-    private var clientOperationalState: ClientOperationalState = .synchronized
-    private var isAutoStarting = false
-    private var isClockSynced = false
+    public internal(set) var currentStreamFormat: AudioFormatSpec?
+    var clientOperationalState: ClientOperationalState = .synchronized
+    var isAutoStarting = false
+    var isClockSynced = false
     /// Incremented on format changes so the scheduler output loop can detect the boundary.
-    private var streamGeneration: UInt64 = 0
+    var streamGeneration: UInt64 = 0
     /// Deferred format + codec header for seamless mid-stream format transitions.
     /// Set in handleStreamStart; consumed by runSchedulerOutput when the first
     /// new-generation chunk arrives.
-    private var pendingFormat: AudioFormatSpec?
-    private var pendingCodecHeader: Data?
+    var pendingFormat: AudioFormatSpec?
+    var pendingCodecHeader: Data?
     /// Current player volume (0-100). Observable for UI binding (volume sliders).
     /// Updated by ``setVolume(_:)`` and by the server via `server/command`.
     public private(set) var currentVolume: Int = 100
@@ -42,39 +46,39 @@ public final class SendspinClient {
     /// Current static delay in milliseconds. Initialized from `PlayerConfiguration.initialStaticDelayMs`,
     /// updated when the server sends a `set_static_delay` command.
     public private(set) var staticDelayMs: Int
-    private var artworkStreamActive = false
+    var artworkStreamActive = false
     /// Cached from `playerConfig?.emitRawAudioEvents` to avoid optional chaining on every audio chunk.
-    private var shouldEmitRawAudio = false
+    var shouldEmitRawAudio = false
 
     /// Accumulated state (merged from server deltas per spec)
     /// Current track metadata, accumulated from `server/state` deltas.
-    public private(set) var currentMetadata: TrackMetadata?
+    public internal(set) var currentMetadata: TrackMetadata?
     /// Current group info, accumulated from `group/update` deltas.
-    public private(set) var currentGroup: GroupInfo?
+    public internal(set) var currentGroup: GroupInfo?
     /// Current controller state from the server.
-    public private(set) var currentControllerState: ControllerState?
+    public internal(set) var currentControllerState: ControllerState?
     /// Codec header for the current stream (e.g. FLAC streaminfo), if any.
     /// Set when `stream/start` carries a `codec_header` field; cleared on `stream/end`.
-    public private(set) var currentCodecHeader: Data?
+    public internal(set) var currentCodecHeader: Data?
 
     // Multi-server state
-    private var currentConnectionReason: ConnectionReason?
-    private var currentServerId: String?
+    var currentConnectionReason: ConnectionReason?
+    var currentServerId: String?
 
     // Dependencies
-    private var transport: (any SendspinTransport)?
-    private var clockSync: ClockSynchronizer?
-    private var audioScheduler: AudioScheduler?
-    private var audioPlayer: AudioPlayer?
+    var transport: (any SendspinTransport)?
+    var clockSync: ClockSynchronizer?
+    var audioScheduler: AudioScheduler?
+    var audioPlayer: AudioPlayer?
 
     // Task management
-    private var messageLoopTask: Task<Void, Never>?
-    private var clockSyncTask: Task<Void, Never>?
-    private var schedulerOutputTask: Task<Void, Never>?
-    private var syncTelemetryTask: Task<Void, Never>?
+    var messageLoopTask: Task<Void, Never>?
+    var clockSyncTask: Task<Void, Never>?
+    var schedulerOutputTask: Task<Void, Never>?
+    var syncTelemetryTask: Task<Void, Never>?
 
     // Event stream
-    private let eventsContinuation: AsyncStream<ClientEvent>.Continuation
+    let eventsContinuation: AsyncStream<ClientEvent>.Continuation
     public let events: AsyncStream<ClientEvent>
 
     public init(
@@ -366,7 +370,7 @@ public final class SendspinClient {
         try await transport.send(ClientHelloMessage(payload: payload))
     }
 
-    private func sendClientState() async throws {
+    func sendClientState() async throws {
         guard let transport else {
             throw SendspinClientError.notConnected
         }
@@ -392,7 +396,7 @@ public final class SendspinClient {
 
     /// Perform initial clock synchronization (5 quick rounds)
     @MainActor
-    private func performInitialSync() async throws {
+    func performInitialSync() async throws {
         guard let transport, let clockSync else {
             throw SendspinClientError.notConnected
         }
@@ -414,7 +418,7 @@ public final class SendspinClient {
         // Otherwise the continuous sync loop will eventually succeed
     }
 
-    private nonisolated func runClockSync() async {
+    nonisolated func runClockSync() async {
         guard let transport = await transport else { return }
 
         while !Task.isCancelled {
@@ -593,429 +597,6 @@ public final class SendspinClient {
         }
     }
 
-    // MARK: - Text message dispatch
-
-    private nonisolated func handleTextMessage(_ text: String) async {
-        let decoder = JSONDecoder()
-        // NOTE: Do NOT use .convertFromSnakeCase — our models define explicit CodingKeys.
-
-        guard let data = text.data(using: .utf8) else { return }
-
-        var msgType = "unknown"
-        if let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-           let type = json["type"] as? String {
-            msgType = type
-            Log.client.debug("RX \(msgType)")
-        }
-
-        // Order matters: messages with required fields before all-optional ones.
-        if let message = try? decoder.decode(ServerHelloMessage.self, from: data), message.type == msgType {
-            await handleServerHello(message)
-        } else if let message = try? decoder.decode(ServerTimeMessage.self, from: data), message.type == msgType {
-            await handleServerTime(message)
-        } else if let message = try? decoder.decode(ServerStateMessage.self, from: data), message.type == msgType {
-            await handleServerState(message)
-        } else if let message = try? decoder.decode(StreamStartMessage.self, from: data), message.type == msgType {
-            await handleStreamStart(message)
-        } else if let message = try? decoder.decode(StreamClearMessage.self, from: data), message.type == msgType {
-            await handleStreamClear(message)
-        } else if let message = try? decoder.decode(StreamEndMessage.self, from: data), message.type == msgType {
-            await handleStreamEnd(message)
-        } else if let message = try? decoder.decode(ServerCommandMessage.self, from: data), message.type == msgType {
-            await handleServerCommand(message)
-        } else if let message = try? decoder.decode(GroupUpdateMessage.self, from: data), message.type == msgType {
-            await handleGroupUpdate(message)
-        } else {
-            let preview = text.prefix(500)
-            Log.client.error("Failed to decode message type '\(msgType)': \(preview)")
-        }
-    }
-
-    // MARK: - Binary message dispatch
-
-    private nonisolated func handleBinaryMessage(_ data: Data) async {
-        guard let message = BinaryMessage(data: data) else { return }
-
-        switch message.type {
-        case .audioChunk:
-            await handleAudioChunkNonisolated(message)
-
-        case .artworkChannel0, .artworkChannel1, .artworkChannel2, .artworkChannel3:
-            await handleArtworkBinary(message)
-
-        case .visualizerData:
-            eventsContinuation.yield(.visualizerData(message.data))
-        }
-    }
-
-    // MARK: - Message handlers
-
-    private func handleServerHello(_ message: ServerHelloMessage) async {
-        connectionState = .connected
-
-        // Track server identity and connection reason for multi-server logic
-        currentServerId = message.payload.serverId
-        currentConnectionReason = message.payload.connectionReason
-
-        let info = ServerInfo(
-            serverId: message.payload.serverId,
-            name: message.payload.name,
-            version: message.payload.version,
-            connectionReason: message.payload.connectionReason,
-            // Wire format is an array; duplicates (if any) are harmlessly deduplicated.
-            activeRoles: Set(message.payload.activeRoles)
-        )
-        eventsContinuation.yield(.serverConnected(info))
-
-        // Send initial client state (required by spec)
-        try? await sendClientState()
-
-        // Perform initial clock sync, then start continuous loop
-        try? await performInitialSync()
-        clockSyncTask = Task.detached { [weak self] in
-            await self?.runClockSync()
-        }
-    }
-
-    private func handleServerTime(_ message: ServerTimeMessage) async {
-        guard let clockSync else { return }
-
-        let now = getCurrentMicroseconds()
-        await clockSync.processServerTime(
-            clientTransmitted: message.payload.clientTransmitted,
-            serverReceived: message.payload.serverReceived,
-            serverTransmitted: message.payload.serverTransmitted,
-            clientReceived: now
-        )
-
-        // Push updated time filter state to the audio callback for sync correction.
-        // This is the only cross-boundary needed — the callback does all the math.
-        // snapshot() returns nil when the filter is uninitialized (before
-        // first sync or after a reset). In that case the audio player keeps
-        // its existing snapshot and plays unsynchronized.
-        if let audioPlayer, let snapshot = await clockSync.snapshot() {
-            await audioPlayer.updateTimeSnapshot(snapshot)
-        }
-    }
-
-    private func handleServerState(_ message: ServerStateMessage) async {
-        // Per spec: "Only include fields that have changed. The client will merge
-        // these updates into existing state. Fields set to null should be cleared."
-        if let metadata = message.payload.metadata {
-            let prev = currentMetadata
-
-            // Build progress: Nullable.merge handles absent vs null vs value
-            let mergedProgress = metadata.progress.merge(previous: prev?.progress.flatMap { p in
-                MetadataProgress(trackProgress: p.trackProgressMs, trackDuration: p.trackDurationMs, playbackSpeed: p.playbackSpeedX1000)
-            })
-            var progress: PlaybackProgress?
-            if let p = mergedProgress {
-                progress = PlaybackProgress(
-                    trackProgressMs: p.trackProgress ?? 0,
-                    trackDurationMs: p.trackDuration ?? 0,
-                    playbackSpeedX1000: p.playbackSpeed ?? 1_000,
-                    timestamp: metadata.timestamp ?? 0
-                )
-            }
-
-            // Merge using Nullable: .absent keeps previous, .null clears, .value updates
-            let merged = TrackMetadata(
-                title: metadata.title.merge(previous: prev?.title),
-                artist: metadata.artist.merge(previous: prev?.artist),
-                album: metadata.album.merge(previous: prev?.album),
-                albumArtist: metadata.albumArtist.merge(previous: prev?.albumArtist),
-                track: metadata.track.merge(previous: prev?.track),
-                year: metadata.year.merge(previous: prev?.year),
-                artworkURL: metadata.artworkUrl.merge(previous: prev?.artworkURL),
-                progress: progress,
-                repeatMode: metadata.repeat.merge(previous: prev?.repeatMode),
-                shuffle: metadata.shuffle.merge(previous: prev?.shuffle)
-            )
-            currentMetadata = merged
-            eventsContinuation.yield(.metadataReceived(merged))
-        }
-
-        // try? is deliberate: throws .notConnected if player role not configured
-        // (see handleServerCommand for rationale)
-        if let player = message.payload.player {
-            if let volume = player.volume {
-                try? await setVolume(volume)
-            }
-            if let muted = player.muted {
-                try? await setMute(muted)
-            }
-        }
-
-        if let controller = message.payload.controller {
-            let prev = currentControllerState
-            let cmds: Set<ControllerCommandType> = controller.supportedCommands
-                .map { Set($0) }
-                ?? prev?.supportedCommands ?? []
-            let vol = min(max(controller.volume ?? prev?.volume ?? 0, 0), 100)
-            let muted = controller.muted ?? prev?.muted ?? false
-
-            let state = ControllerState(
-                supportedCommands: cmds,
-                volume: vol,
-                muted: muted
-            )
-            currentControllerState = state
-            eventsContinuation.yield(.controllerStateUpdated(state))
-        }
-    }
-
-    private func handleStreamStart(_ message: StreamStartMessage) async {
-        // Handle artwork stream start
-        if let artworkInfo = message.payload.artwork {
-            artworkStreamActive = true
-            eventsContinuation.yield(.artworkStreamStarted(artworkInfo.channels))
-        }
-
-        // Handle player stream start
-        guard let playerInfo = message.payload.player else {
-            Log.client.info("stream/start: artwork only (no player payload)")
-            return
-        }
-        guard let audioPlayer else {
-            Log.client.error("stream/start: player payload received but audioPlayer is nil!")
-            return
-        }
-        Log.client.info("stream/start: \(playerInfo.codec) \(playerInfo.sampleRate)Hz \(playerInfo.channels)ch \(playerInfo.bitDepth)bit")
-
-        guard let codec = AudioCodec(rawValue: playerInfo.codec) else {
-            connectionState = .error(.unsupportedCodec(playerInfo.codec))
-            clientOperationalState = .error
-            try? await sendClientState()
-            return
-        }
-
-        let format = AudioFormatSpec(
-            codec: codec,
-            channels: playerInfo.channels,
-            sampleRate: playerInfo.sampleRate,
-            bitDepth: playerInfo.bitDepth
-        )
-
-        var codecHeader: Data?
-        if let headerBase64 = playerInfo.codecHeader {
-            codecHeader = Data(base64Encoded: headerBase64)
-        }
-        currentCodecHeader = codecHeader
-        shouldEmitRawAudio = playerConfig?.emitRawAudioEvents ?? false
-
-        let wasPlaying = await audioPlayer.isPlaying
-        let previousFormat = currentStreamFormat
-        let isFormatChange = wasPlaying && previousFormat != nil && previousFormat != format
-
-        if isFormatChange {
-            // Seamless mid-stream format transition.
-            //
-            // 1. Bump generation — chunks are tagged at scheduling time, so
-            //    runSchedulerOutput can tell old-format from new-format chunks
-            // 2. Swap decoder — new binary chunks get decoded at the new rate
-            // 3. Store pending format for deferred AudioQueue rebuild
-            // 4. DON'T clear scheduler — old chunks continue playing through
-            //    the old AudioQueue, providing seamless audio coverage
-            // 5. DON'T stop AudioQueue — it keeps running from its ring buffer
-            //
-            // When runSchedulerOutput sees the first chunk with the new generation,
-            // it rebuilds the AudioQueue at the new sample rate. Old-gen chunks
-            // play through the old AudioQueue until that point.
-            streamGeneration &+= 1
-            pendingFormat = format
-            pendingCodecHeader = codecHeader
-            currentStreamFormat = format
-
-            // Swap decoder for new incoming chunks
-            do {
-                try await audioPlayer.swapDecoder(format: format, codecHeader: codecHeader)
-            } catch {
-                Log.client.error("Decoder swap failed, full restart: \(error)")
-                pendingFormat = nil
-                pendingCodecHeader = nil
-                try? await audioPlayer.start(format: format, codecHeader: codecHeader)
-            }
-
-            let oldRate = previousFormat!.sampleRate
-            Log.client.info("Format change: \(oldRate)Hz → \(format.sampleRate)Hz")
-            eventsContinuation.yield(.streamFormatChanged(format))
-            try? await sendClientState()
-            return
-        }
-
-        do {
-            try await audioPlayer.start(format: format, codecHeader: codecHeader)
-            currentStreamFormat = format
-            clientOperationalState = .synchronized
-            await audioScheduler?.startScheduling()
-
-            if !wasPlaying {
-                eventsContinuation.yield(.streamStarted(format))
-            }
-            try? await sendClientState()
-        } catch {
-            connectionState = .error(.audioStartFailed(error.localizedDescription))
-            clientOperationalState = .error
-            try? await sendClientState()
-        }
-    }
-
-    private func handleStreamClear(_ message: StreamClearMessage) async {
-        let roles = message.payload.roles
-
-        if roles == nil || roles?.contains("player") == true {
-            await audioScheduler?.clear()
-            if let audioPlayer {
-                await audioPlayer.clearBuffer()
-            }
-            eventsContinuation.yield(.streamCleared)
-        }
-    }
-
-    private func handleServerCommand(_ message: ServerCommandMessage) async {
-        guard let playerCmd = message.payload.player else { return }
-
-        // try? is deliberate: these throw .notConnected if audioPlayer is nil
-        // (player role not configured), but the server may send player commands
-        // to any client. Silently ignoring commands for unconfigured roles is
-        // correct — the server doesn't know our role configuration until it
-        // reads our client/hello.
-        switch playerCmd.command {
-        case .volume:
-            if let volume = playerCmd.volume {
-                try? await setVolume(volume)
-            }
-        case .mute:
-            if let mute = playerCmd.mute {
-                try? await setMute(mute)
-            }
-        case .setStaticDelay:
-            if let delayMs = playerCmd.staticDelayMs {
-                try? await setStaticDelay(delayMs)
-            }
-        }
-    }
-
-    private func handleStreamEnd(_ message: StreamEndMessage) async {
-        // stream/end with no roles ends all streams
-        let endedRoles = message.payload.roles
-
-        if endedRoles == nil || endedRoles?.contains("player") == true {
-            if let audioPlayer {
-                await audioScheduler?.stop()
-                await audioScheduler?.clear()
-                await audioPlayer.stop()
-            }
-            currentStreamFormat = nil
-            currentCodecHeader = nil
-        }
-
-        if endedRoles == nil || endedRoles?.contains("artwork") == true {
-            artworkStreamActive = false
-        }
-
-        clientOperationalState = .synchronized
-        eventsContinuation.yield(.streamEnded)
-    }
-
-    private func handleGroupUpdate(_ message: GroupUpdateMessage) async {
-        // Per spec: persist server_id when playback_state transitions to 'playing'.
-        // Emitted as an event so the host app can persist using its own storage.
-        if message.payload.playbackState == .playing, let serverId = currentServerId {
-            eventsContinuation.yield(.lastPlayedServerChanged(serverId: serverId))
-        }
-
-        // Per spec: "Contains delta updates with only the changed fields.
-        // The client should merge these updates into existing state."
-        let prev = currentGroup
-        let playbackState = message.payload.playbackState ?? prev?.playbackState
-
-        let info = GroupInfo(
-            groupId: message.payload.groupId ?? prev?.groupId ?? "",
-            groupName: message.payload.groupName ?? prev?.groupName ?? "",
-            playbackState: playbackState
-        )
-        currentGroup = info
-        eventsContinuation.yield(.groupUpdated(info))
-    }
-
-    // MARK: - Audio chunk handling
-
-    private nonisolated func handleAudioChunkNonisolated(_ message: BinaryMessage) async {
-        await handleAudioChunk(message)
-    }
-
-    /// Handle artwork binary frames on the main actor.
-    ///
-    /// The spec says binary messages "should be rejected if there is no active stream."
-    /// We intentionally don't gate on `artworkStreamActive` here because binary and text
-    /// messages arrive on parallel tasks — artwork binary frames can (and do) arrive
-    /// before the `stream/start` text message that sets the flag. The binary message
-    /// type (8-11) already validates this is artwork data from the server, which is
-    /// sufficient validation. Dropping legitimate frames due to a race would be worse
-    /// than delivering them slightly early.
-    private func handleArtworkBinary(_ message: BinaryMessage) {
-        guard let channel = message.type.artworkChannel else { return }
-        // Empty payload (no image data) means clear the artwork per spec
-        eventsContinuation.yield(.artworkReceived(channel: channel, data: message.data))
-    }
-
-    private func handleAudioChunk(_ message: BinaryMessage) async {
-        // Emit raw audio data before any processing for conformance testing / recording.
-        // This must be above the clock-sync guard — the conformance adapter needs every
-        // chunk regardless of sync state.
-        if shouldEmitRawAudio {
-            eventsContinuation.yield(.rawAudioChunk(data: message.data, serverTimestamp: message.timestamp))
-        }
-
-        // Don't process audio until clock sync is complete — timestamps would be wrong
-        if !isClockSynced {
-            if let clockSync, await clockSync.hasSynced {
-                isClockSynced = true
-                await audioScheduler?.clear()
-            } else {
-                return
-            }
-        }
-
-        guard let audioPlayer,
-              let audioScheduler
-        else { return }
-
-        // Auto-start player if not already started (some servers don't send stream/start)
-        let isPlaying = await audioPlayer.isPlaying
-        if !isPlaying, !isAutoStarting {
-            isAutoStarting = true
-
-            guard let defaultFormat = playerConfig?.supportedFormats.first else {
-                isAutoStarting = false
-                return
-            }
-
-            do {
-                try await audioPlayer.start(format: defaultFormat, codecHeader: nil)
-                clientOperationalState = .synchronized
-                await audioScheduler.startScheduling()
-                eventsContinuation.yield(.streamStarted(defaultFormat))
-                try? await sendClientState()
-            } catch {
-                isAutoStarting = false
-                return
-            }
-        } else if !isPlaying, isAutoStarting {
-            return // Another chunk is already triggering auto-start
-        }
-
-        do {
-            let pcmData = try await audioPlayer.decode(message.data)
-            // Per spec: subtract static_delay_ms from server timestamp before scheduling
-            let adjustedTimestamp = message.timestamp - Int64(staticDelayMs) * 1_000
-            await audioScheduler.schedule(pcm: pcmData, serverTimestamp: adjustedTimestamp, generation: streamGeneration)
-        } catch {
-            // Decode/schedule failure — drop this chunk
-        }
-    }
-
     // MARK: - Utilities
 
     /// Send a message through the transport, wrapping any transport error in
@@ -1031,7 +612,7 @@ public final class SendspinClient {
         }
     }
 
-    private nonisolated func getCurrentMicroseconds() -> Int64 {
+    nonisolated func getCurrentMicroseconds() -> Int64 {
         MonotonicClock.nowMicroseconds()
     }
